# feat: deduplicate events emision in nested properties