--- a/src/psygnal/_evented_model_v1.py
+++ b/src/psygnal/_evented_model_v1.py
@@ -1,6 +1,7 @@
 import sys
 import warnings
 from contextlib import contextmanager
+from types import TracebackType
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -8,6 +9,7 @@ from typing import (
     ClassVar,
     Dict,
     Iterator,
+    Optional,
     Set,
     Type,
     Union,
@@ -15,7 +17,7 @@ from typing import (
     no_type_check,
 )
 
-from ._evented_model_utils import ComparisonDelayer
+# from ._evented_model_utils import ComparisonDelayer
 from ._group import SignalGroup
 from ._group_descriptor import _check_field_equality, _pick_equality_operator
 from ._signal import Signal, SignalInstance
@@ -522,3 +524,20 @@ def _get_defaults(obj: BaseModel) -> Dict[str, Any]:
             d = _get_defaults(v.type_)  # pragma: no cover
         dflt[k] = d
     return dflt
+
+
+class ComparisonDelayer:
+    def __init__(self, target: EventedModel) -> None:
+        self._target = target
+
+    def __enter__(self) -> None:
+        self._target._delay_check_semaphore += 1
+
+    def __exit__(
+        self,
+        exc_type: Optional[type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> None:
+        self._target._delay_check_semaphore -= 1
+        self._target._check_if_values_changed_and_emit_if_needed()