This PR migrates https://github.com/napari/napari/pull/6075 into psygnal to deduplicated event emission.

This will allow in future to drop the `napari.utils.events` module.