@tlambert03 2024-02-23
thanks @Czaki, would be happy to do what is needed here to let napari drop the internal event object.

looks like this has a significant negative impact on the time for setattr in EventedModel, but with and without connections

| Change   | Before [9cdce45f] <main>   | After [8c62e562]    |   Ratio | Benchmark (Parameter)                                           |
|----------|----------------------------|---------------------|---------|-----------------------------------------------------------------|
| +        | 815±4μs                    | 4.42±0.03ms         |    5.42 | benchmarks.EventedModelSuite.time_setattr_with_connections(100) |
| +        | 28.2±0.09μs                | 134±2μs             |    4.76 | benchmarks.EventedModelSuite.time_setattr_no_connections(10)    |
| +        | 276±2μs                    | 1.30±0.01ms         |    4.73 | benchmarks.EventedModelSuite.time_setattr_no_connections(100)   |
| +        | 210±2μs                    | 509±4μs             |    2.42 | benchmarks.EventedModelSuite.time_setattr_with_connections(10)  |
| +        | 68.2±0.4ms                 | 83.7±10ms           |    1.23 | benchmarks.EventedSetSuite.time_update_new(100000)              |


is the ComparisonDelayer fundamental to the deduplication? or can we look at adding that in a separate PR?