@tlambert03 2024-02-23
(also, feel free to add benchmarks to the suite that show what is *gained* from this PR, so we can balance our understanding of what is lost)