@tlambert03 2024-02-23
one more question: is the problem you were trying to solve in #6075 related only to events on computed properties?  If I run the test you've added here on main, it passes except for the very last line: `e_m.assert_called_once()` ... which is called twice.

If so, I'd like it if we could step back and discuss a broader question here first.  I have to say i think the whole ability to add property-setters to a pydantic model and to calculate events on them may well have been an over-implementation on my part to begin with.  It's a whole new thing that pydantic didn't support at the time, and now *does*  seem to be [supporting in v2](https://docs.pydantic.dev/2.0/usage/computed_fields/).  So, if anything, I think we should perhaps be deprecating and de-emphasizing this little feature in psygnal, instead of elaborating on it at the potential expense of performance for non-property-setting classes.

I do definitely want to support your usage of EventedModel over in napari, but i'd like to understand whether this is actually a good broader design decision here, or if we might just be "inheriting" complicated patterns from napari.

@andy-sweet, I'd be grateful for your input here as well