@tlambert03 2024-02-23
> The general performance gain is to avoid heavy callbacks execution (for example, update gui).

Yep that makes sense, but as the benchmarks here are suggesting, it's possible that the performance gain for properties comes at the expense of performance of the (more important) regular fields.

So, we might need this to be an opt-in feature that only applies if someone is actively trying to use this relatively uncommon pattern