@Czaki 2024-02-23
ok. The test was wrongly written. It checks multiple cases, which here are covered by other tests. So it could be simplified.