@tlambert03 2024-02-23
this is great! Thanks @Czaki!  I like the selection of setattr in `__new__`.  good solution.

I'm inclined to merge as is.  anything else you want to do here?